uv-pypi-types: rename conflicting_groups module to conflicts
We'll do the rest of the renaming in the next commit.
