+++ crates/uv-pypi-types/src/lib.rs
@@ -1,5 +1,5 @@
 pub use base_url::*;
-pub use conflicting_groups::*;
+pub use conflicts::*;
 pub use direct_url::*;
 pub use lenient_requirement::*;
 pub use marker_environment::*;
@@ -11,7 +11,7 @@ pub use simple_json::*;
 pub use supported_environments::*;
 
 mod base_url;
-mod conflicting_groups;
+mod conflicts;
 mod direct_url;
 mod lenient_requirement;
 mod marker_environment;
